feat(ml):In the ML module, a new 'methods' folder has been added, where deep learning method files can be added in parallel according to plan. These files encompass the core functionalities or algorithms of their respective methods. In the newly added PINN Python file, a residual class Residual is defined. An example of its usage in the training script is: residual = Residual('helmholtz_2d'), where the PDE model name is passed as an instantiation parameter. This allows for returning the corresponding residual function for different PDE models, achieving decoupling of methods and models from the user's perspective and offering excellent scalability.
